float -> qreal. Hope this doesn't break for anyone
[git-p4: depot-paths = "//depot/qt/main/": change = 168517]

+++ src/core/tools/qrect.cpp
@@ -1480,7 +1480,7 @@ bool QRectF::intersects(const QRectF &r) const
 
 QDataStream &operator<<(QDataStream &s, const QRectF &r)
 {
-    s << r.x() << r.y() << r.width() << r.height();
+    s << (double)r.x() << (double)r.y() << (double)r.width() << (double) r.height();
     return s;
 }
 
@@ -1497,7 +1497,7 @@ QDataStream &operator<<(QDataStream &s, const QRectF &r)
 
 QDataStream &operator>>(QDataStream &s, QRectF &r)
 {
-    float x, y, w, h;
+    double x, y, w, h;
     s >> x; s >> y; s >> w; s >> h;
     r.setRect(x, y, w, h);
     return s;
